Rename: 날짜 변환 함수 utils로 분리
